pyproject.toml
[project]
name = "longevity_forest"
version = "0.1.1"
description = "Multi-agent bioinformatics research system for agentic analysis focusing on longevity"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "just-agents>=0.8.8",
    "typer>=0.20.0",
    "win-unicode-console>=0.5", #for poor windows users
    "pyyaml>=6.0.3",
    "pycomfort>=0.0.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/longevity_forest"]

[project.scripts]
longevity_forest = "longevity_forest.main:main"
forest = "longevity_forest.main:app"
